DB test 로직 추가 및 수정

+++ model/message.go
@@ -2,7 +2,7 @@ package model
 
 // Message 커뮤니케이션용 객체
 type Message struct {
-	ID      int    `json:"id,omitempty"`
+	ID      int64  `json:"id,omitempty"`
 	Sender  string `json:"sender"`
 	Message string `json:"message"`
 }
